requirements.txt
absl-py==0.11.0
aiohttp==3.6.3
apache-beam==2.27.0
appdirs==1.4.4
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
autograd==1.3
avro-python3==1.9.2.1
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
brotlipy @ file:///D:/bld/brotlipy_1636012381660/work
cachetools==4.1.1
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1606255208697/work
chardet==3.0.4
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1638815705608/work
chart-studio==1.1.0
cityscapesScripts==2.2.0
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
cma==2.7.0
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
coloredlogs==15.0.1
contextlib2==0.6.0.post1
crcmod==1.7
cryptography @ file:///D:/bld/cryptography_1639699508385/work
cycler==0.10.0
Cython==0.29.27
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
efemarai==1.3.0
efficientnet==1.0.0
entrypoints==0.3
fastavro==1.2.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.6.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
hdfs==2.5.8
humanfriendly==10.0
hyperopt==0.1.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
image-classifiers==1.0.0
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
importlib-resources==5.1.2
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1604083276484/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///C:/ci/jedi_1606914528444/work
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
json-tricks==3.15.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///C:/ci/jupyter_core_1606149073702/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kaggle==1.5.10
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1604014703538/work
lvis==0.5.3
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1603355780617/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==2.0.0
multidict==4.7.6
nas-bench-201==2.1
-e git+https://github.com/google-research/nasbench@b94247037ee470418a3e56dcb83814e9be83f3a8#egg=nasbench
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
netifaces==0.10.9
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
nni==2.0
notebook @ file:///C:/ci/notebook_1601501643625/work
numpy==1.19.3
oauth2client==4.1.3
oauthlib==3.1.0
object-detection @ file:///C:/Users/Daniel/Documents/Development/Object_detection/models/research
olefile==0.46
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1606930849755/work
pandas @ file:///C:/ci/pandas_1602083338010/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso==0.7.0
persist-queue==0.5.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1603823068645/work
-e git+https://github.com/Project-Platypus/Platypus.git@a4e56410a772798e905407e99f80d03b86296ad3#egg=Platypus_Opt
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1640064922437/work
plyer==2.0.0
prettytable==2.0.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.15.8
psutil==5.8.0
py-cpuinfo==7.0.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycocotools-windows==2.0.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydot==1.4.1
pygad==2.16.1
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pymongo==3.11.2
pymoo==0.5.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==2.4.7
pyquaternion==0.9.9
pyreadline3==3.4.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///D:/bld/pysocks_1635862590164/work
python-dateutil==2.8.1
python-engineio==3.14.2
python-socketio==4.6.1
PythonWebHDFS==0.2.3
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1637771257551/work
requests-oauthlib==1.3.0
responses==0.12.1
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.4
scikit-image==0.18.3
scikit-learn @ file:///C:/ci/scikit-learn_1598377018496/work
scikit-multilearn==0.2.0
scipy @ file:///C:/ci/scipy_1604596260408/work
seaborn==0.11.1
segmentation-models==1.0.1
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==1.2.2
simplejson==3.17.2
sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
tb-nightly==2.5.0a20201203
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1626090218611/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@228643b2062756ce2c156e295adfff5b530a0ae5
tensorflow-metadata==0.29.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2020120301
tf-models-official==2.4.0
tf-nightly-gpu==2.5.0.dev20201203
tf-slim==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2021.10.12
torch==1.8.1
torchvision==0.9.1
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///tmp/build/80754af9/tqdm_1606938474023/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
typeguard==2.13.0
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1632350318291/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1635833007271/work
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
